Refactored A* algorithm references to 'A Star'
The code changes involve renaming all instances of 'A*' in file names, import statements, and text content to 'A Star'. This includes updates in component imports, style imports, enum references, and portfolio content. The refactoring ensures consistency in naming across the application.
