+++ src/components/algorithms/aStar/aStarNode.ts
