+++ src/components/algorithms/aStar/styles/aStarStyleMap.module.scss
