+++ src/components/algorithms/aStar/utils/aStarResult.ts
@@ -1,6 +1,6 @@
 import { GridState } from "../../grid/gridState";
-import { AStarStages } from "./a*stages.enum";
-import { AStarStates } from "./a*states.enum";
+import { AStarStages } from "./aStarStages.enum";
+import { AStarStates } from "./aStarStates.enum";
 
 
 export class AStarResult {
