+++ src/components/algorithms/aStar/utils/aStarStates.enum.ts
